+++ src/handle_mousedown.rs
@@ -3,7 +3,7 @@ use crate::{touches::Touches, Msg};
 pub fn handle_mousedown(
     x: i32,
     y: i32,
-    current_ts: u128,
+    current_ts: u64,
     touches: &mut Touches,
     is_right_click: bool,
 ) -> Msg {
